Add FGI cache. Enhance quote cache and inspection scripts with improved path handling and error logging
- Update QuoteCache initialization to resolve relative cache directory paths against the project root
- Modify inspect_cache script to handle relative paths and improve error messages for module imports and cache directory checks
- Standardize error and warning messages for better clarity and consistency
- Update backtest configuration to switch data sources and adjust parameters for testing

+++ core/models/fill.py
@@ -5,7 +5,7 @@ from datetime import datetime
 class Fill:
     order_id: str
     symbol: str
-    quantity: int
+    quantity: float  # Changed to float to support fractional shares
     price: float
     commission: float
     timestamp: datetime
